Create fundamentals.py
I think that functions are probably some of the more difficult things to understand for beginners. Just keep in mind that the programs that you might want to develop or even the ones that you are currently using are comprised of many functions. When you break open and look at the code for some simple programs you will see lots of built in functions and user defined functions being called.

I say all of this to help drive home the fact that even though this might be a tough spot for you learning progression, its one of the most crucial parts and key component of the fundamentals.